requirements.txt


asgiref==3.6.0
Django==4.1.7
django-rest-framework-mongoengine==3.4.1
djangorestframework==3.14.0
djongo==1.3.6
dnspython==2.3.0
mongoengine==0.26.0
pip==22.3.1
pymongo==3.12.1
pytz==2022.7.1
setuptools==66.1.1
sqlparse==0.2.4
tzdata==2022.7
wheel==.38.4
